Remover botões copiar, colar, recortar e adicionar botão de syscalls na toolbar
