+++ mars/venus/VenusUI.java
@@ -76,13 +76,13 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
               settingsExtended, settingsAssembleOnOpen, settingsAssembleAll, settingsWarningsAreErrors, settingsStartAtMain,
       		  settingsDelayedBranching, settingsProgramArguments, settingsSelfModifyingCode;
       private JMenuItem settingsExceptionHandler, settingsEditor, settingsHighlighting, settingsMemoryConfiguration;
-      private JMenuItem helpHelp, helpAbout;
+      private JMenuItem helpHelp, helpSyscalls, helpAbout;
          
       // components of the toolbar
-      private JButton Undo, Redo, Cut, Copy, Paste, FindReplace, SelectAll;
+      private JButton Undo, Redo, FindReplace, SelectAll;
       private JButton New, Open, Save, SaveAs, SaveAll, DumpMemory, Print;
       private JButton Run, Assemble, Reset, Step, Backstep, Stop, Pause;
-      private JButton Help;
+      private JButton Help, HelpSyscalls;
    
       // The "action" objects, which include action listeners.  One of each will be created then
    	// shared between a menu item and its corresponding toolbar button.  This is a very cool
@@ -100,7 +100,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
       					settingsWarningsAreErrorsAction, settingsStartAtMainAction, settingsProgramArgumentsAction,
       					settingsDelayedBranchingAction, settingsExceptionHandlerAction, settingsEditorAction,
       					settingsHighlightingAction, settingsMemoryConfigurationAction, settingsSelfModifyingCodeAction;    
-      private Action helpHelpAction, helpAboutAction;
+      private Action helpHelpAction, helpSyscallsAction, helpAboutAction;
    
    
     /**
@@ -446,7 +446,13 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
                                             new ImageIcon(tk.getImage(cs.getResource(Globals.imagesPath+"Help22.png"))),
                									  "Help", new Integer(KeyEvent.VK_H),
                									  KeyStroke.getKeyStroke(KeyEvent.VK_F1, 0),
-               									  mainUI);	
+               									  mainUI);
+            helpSyscallsAction = new HelpSyscallsAction("Syscalls",
+                                            new ImageIcon(tk.getImage(cs.getResource(Globals.imagesPath + "Syscall22.png"))),
+                                            "Help on Syscalls",
+                                            KeyEvent.VK_S,
+                                            null,
+                                            mainUI);
             helpAboutAction = new HelpAboutAction("About ...",null, 
                                             "Information about Mars", null,null, mainUI);	
          } 
@@ -626,9 +632,14 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
       			
          helpHelp = new JMenuItem(helpHelpAction);
          helpHelp.setIcon(new ImageIcon(tk.getImage(cs.getResource(Globals.imagesPath+"Help16.png"))));//"Help16.gif"))));
+
+         helpSyscalls = new JMenuItem(helpSyscallsAction);
+         helpSyscalls.setIcon(new ImageIcon(tk.getImage(cs.getResource(Globals.imagesPath+"Syscall16.png"))));
+
          helpAbout = new JMenuItem(helpAboutAction);
          helpAbout.setIcon(new ImageIcon(tk.getImage(cs.getResource(Globals.imagesPath+"MyBlank16.gif"))));
          help.add(helpHelp);
+         help.add(helpSyscalls);
          help.addSeparator();
          help.add(helpAbout);
       
@@ -671,12 +682,6 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
          Undo.setText(""); 
          Redo = new JButton(editRedoAction);
          Redo.setText("");   	
-         Cut= new JButton(editCutAction);
-         Cut.setText("");
-         Copy= new JButton(editCopyAction);
-         Copy.setText("");
-         Paste= new JButton(editPasteAction);
-         Paste.setText("");
          FindReplace = new JButton(editFindReplaceAction);
          FindReplace.setText("");
          SelectAll = new JButton(editSelectAllAction);
@@ -698,6 +703,8 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
          Pause.setText("");      	
          Help= new JButton(helpHelpAction);
          Help.setText("");
+         HelpSyscalls = new JButton(helpSyscallsAction);
+         HelpSyscalls.setText("");
          
          toolBar.add(New);
          toolBar.add(Open);
@@ -710,9 +717,6 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
          toolBar.add(new JToolBar.Separator());
          toolBar.add(Undo);
          toolBar.add(Redo);
-         toolBar.add(Cut);
-         toolBar.add(Copy);
-         toolBar.add(Paste);
          toolBar.add(FindReplace);
          toolBar.add(new JToolBar.Separator());
          toolBar.add(Assemble);
@@ -724,6 +728,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
          toolBar.add(Reset);
          toolBar.add(new JToolBar.Separator());
          toolBar.add(Help);
+         toolBar.add(HelpSyscalls);
          toolBar.add(new JToolBar.Separator());
       	
          return toolBar;
@@ -806,6 +811,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
          runClearBreakpointsAction.setEnabled(false);
          runToggleBreakpointsAction.setEnabled(false);
          helpHelpAction.setEnabled(true);
+         helpSyscallsAction.setEnabled(true);
          helpAboutAction.setEnabled(true);
          editUndoAction.updateUndoState();
          editRedoAction.updateRedoState();
@@ -847,6 +853,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
             runToggleBreakpointsAction.setEnabled(false);
          } 
          helpHelpAction.setEnabled(true);
+         helpSyscallsAction.setEnabled(true);
          helpAboutAction.setEnabled(true);
          editUndoAction.updateUndoState();
          editRedoAction.updateRedoState();
@@ -884,6 +891,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
          runClearBreakpointsAction.setEnabled(false);
          runToggleBreakpointsAction.setEnabled(false);
          helpHelpAction.setEnabled(true);
+         helpSyscallsAction.setEnabled(true);
          helpAboutAction.setEnabled(true);
          editUndoAction.updateUndoState();
          editRedoAction.updateRedoState();
@@ -920,6 +928,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
          runClearBreakpointsAction.setEnabled(false);
          runToggleBreakpointsAction.setEnabled(false);
          helpHelpAction.setEnabled(true);
+         helpSyscallsAction.setEnabled(true);
          helpAboutAction.setEnabled(true);
          editUndoAction.updateUndoState();
          editRedoAction.updateRedoState();
@@ -957,6 +966,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
          runPauseAction.setEnabled(false);
          runToggleBreakpointsAction.setEnabled(true);
          helpHelpAction.setEnabled(true);
+         helpSyscallsAction.setEnabled(true);
          helpAboutAction.setEnabled(true);
          editUndoAction.updateUndoState();
          editRedoAction.updateRedoState();
@@ -992,6 +1002,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
          runPauseAction.setEnabled(true);
          runToggleBreakpointsAction.setEnabled(false);
          helpHelpAction.setEnabled(true);
+         helpSyscallsAction.setEnabled(true);
          helpAboutAction.setEnabled(true);
          editUndoAction.setEnabled(false);//updateUndoState(); // DPS 10 Jan 2008
          editRedoAction.setEnabled(false);//updateRedoState(); // DPS 10 Jan 2008
@@ -1028,6 +1039,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
          runPauseAction.setEnabled(false);
          runToggleBreakpointsAction.setEnabled(true);
          helpHelpAction.setEnabled(true);
+         helpSyscallsAction.setEnabled(true);
          helpAboutAction.setEnabled(true);
          editUndoAction.updateUndoState();
          editRedoAction.updateRedoState();
